exercise_1_simple-rag-system/requirements.txt
# Embeddings
sentence-transformers>=2.2.0

# Vector store
numpy>=1.24.0

# LLM
openai>=1.0.0

# Tokenization
tiktoken>=0.5.0

# Environment variables
python-dotenv>=1.0.0